Sort into agent module
